The view is coming along nicely. I had to do some complex maneuverings to figure out if they have enough points since we stored values in strings, but need to convert them to numbers to do comparisons. Despite this, I have hammered out the most important elements of the view and the logic works great (see attached).
I still need to build out the mobile view. Even though you want it to look the same, I need to code a separate view on account of the sidenav on desktop. It's not hard, just time consuming.

I also have to add the drop down when people hit Overview Details, but that's easy.

As a result, it will definitely be done by Monday, and I should have tackled the major bugs by COB Monday as well, barring any new ones you find related to UC-05.